refactor Colloboque server, SQL generation moved to standalone file
